+++ src/index.ts
@@ -1,10 +1,9 @@
 import { Connection } from "./Connection";
-import { ConnectionConfig } from "./types";
+import { ConnectionConfig, StatisticsPacket, EventNames, ChangeUserOptions } from "./types";
 import { QueryBuilder } from "./QueryBuilder";
 
 export async function createConnection(config: ConnectionConfig): Promise<Connection> {
   return new Connection(config);
 }
 
-export { Connection, QueryBuilder };
-export type { ConnectionConfig };
+export { Connection, QueryBuilder, ConnectionConfig, StatisticsPacket, EventNames, ChangeUserOptions };