+++ src/types.ts
@@ -3,3 +3,38 @@ import mysql from "mysql";
 export interface ConnectionConfig extends mysql.ConnectionConfig {
   reconnect?: boolean;
 }
+
+export interface StatisticsPacket {
+  message: string;
+  uptime: number;
+  threads: number;
+  questions: number;
+  slow_queries: number;
+  opens: number;
+  flush_tables: number;
+  queries_per_second_avg: number;
+}
+
+export interface ChangeUserOptions {
+  /**
+   * The name of the new user (defaults to the previous one)
+   */
+  user?: string;
+
+  /**
+   * The password of the new user (defaults to the previous one)
+   */
+  password?: string;
+
+  /**
+   * The new charset (defaults to the previous one)
+   */
+  charset?: string;
+
+  /**
+   * The new database (defaults to the previous one)
+   */
+  database?: string;
+}
+
+export type EventNames = "error" | "end";